requirements.txt
bcrypt==3.1.7
bootstrap3-datetime==2.4
boto==2.49.0
cffi==1.14.0
cryptography==2.8
Django==2.1.1
django-autocomplete-light==3.5.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap3-datepicker==0.4.1
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-jquery==3.1.0
django-qr-code==1.0.0
fabric==2.5.0
humanize==0.5.1
invoke==1.4.1
paramiko==2.7.1
Pillow==5.2.0
prettytable==0.7.2
pycparser==2.20
PyNaCl==1.3.0
python-barcode==0.10.0
pytz==2018.5
qrcode==6.1
reportlab==3.5.23
six==1.12.0
treepoem==3.0.0
